engine: Add the -r option to resume training from an existing file.
TODO: We might want to modify the CLI a bit more to store data about
which datasets we are using for the training as well as the current
convergence factor and max iterations. The advantages of this are that
we have more information for reference when looking at the datasets in
general but also so that we can simply do something like
"python engine.py -r filename.pkl" to resume training instead of
figuring out what parameters we used during the initial training.